refactor: PostAddDialog UI 분리
