+++ src/feature/post/ui/PostAddDialog.tsx
@@ -1,0 +1,49 @@
+import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Input, Textarea } from "../../../shared/ui"
+import { useAddPostMutation, useDialogStore, useNewPostStore } from "../model"
+
+export const PostAddDialog = () => {
+  const { newPost, setNewPost, resetNewPost } = useNewPostStore()
+  const { showAddDialog, setShowAddDialog } = useDialogStore()
+
+  const addPostMutation = useAddPostMutation()
+
+  const handleAddPost = async () => {
+    try {
+      await addPostMutation.mutateAsync(newPost)
+      setShowAddDialog(false)
+      resetNewPost()
+    } catch (error) {
+      console.error("게시물 추가 오류:", error)
+    }
+  }
+
+  return (
+    <Dialog open={showAddDialog} onOpenChange={setShowAddDialog}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>새 게시물 추가</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <Input
+            placeholder="제목"
+            value={newPost.title}
+            onChange={(e) => setNewPost({ ...newPost, title: e.target.value })}
+          />
+          <Textarea
+            rows={30}
+            placeholder="내용"
+            value={newPost.body}
+            onChange={(e) => setNewPost({ ...newPost, body: e.target.value })}
+          />
+          <Input
+            type="number"
+            placeholder="사용자 ID"
+            value={newPost.userId}
+            onChange={(e) => setNewPost({ ...newPost, userId: Number(e.target.value) })}
+          />
+          <Button onClick={handleAddPost}>게시물 추가</Button>
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
