+++ src/feature/post/ui/index.ts
@@ -2,3 +2,4 @@ export * from "./PostTitleCell"
 export * from "./PostDetailButton"
 export * from "./PostEditButton"
 export * from "./PostDeleteButton"
+export * from "./PostAddDialog"
